+++ src/components/render.js
@@ -1,0 +1,5 @@
+const render = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export {render};
